feat(injector): add ephemeral-storage resource request/limit for sidecars
