+++ pkg/plugins/runtime/k8s/webhooks/injector/testdata/inject.06.golden.yaml
@@ -98,9 +98,11 @@ spec:
       limits:
         cpu: 1100m
         memory: 1512Mi
+        ephemeral-storage: 10M
       requests:
         cpu: 150m
         memory: 164Mi
+        ephemeral-storage: 10M
     securityContext:
       readOnlyRootFilesystem: true
       runAsGroup: 5678
