+++ pkg/plugins/runtime/k8s/webhooks/injector/testdata/inject.08.golden.yaml
@@ -100,9 +100,11 @@ spec:
       limits:
         cpu: 1100m
         memory: 1512Mi
+        ephemeral-storage: 10M
       requests:
         cpu: 150m
         memory: 164Mi
+        ephemeral-storage: 10M
     securityContext:
       readOnlyRootFilesystem: true
       runAsGroup: 5678
