+++ pkg/plugins/runtime/k8s/webhooks/injector/testdata/inject.28.golden.yaml
@@ -102,9 +102,11 @@ spec:
       limits:
         cpu: 1100m
         memory: 1512Mi
+        ephemeral-storage: 10M
       requests:
         cpu: 150m
         memory: 164Mi
+        ephemeral-storage: 10M
     securityContext:
       privileged: false
       readOnlyRootFilesystem: true
